% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiple_stress}
\alias{multiple_stress}
\title{Survival of Daphnia magna exposed to multiple stressors}
\format{
A data frame with 58 rows and 4 variables:
\describe{
  \item{food}{The amount of food in percent.}
  \item{esfenvalerate}{The concentration of esfenvalerate in µg/L.}
  \item{prochloraz}{The concentration of prochloraz in µg/L.}
  \item{survival}{The mean survival in percent.}
}
}
\source{
\href{https://doi.org/10.1021/acs.est.9b04293}{Shahid, N., Liess, M.,
    Knillmann, S., 2019. Environmental Stress Increases Synergistic Effects
    of Pesticide Mixtures on Daphnia magna. Environ. Sci. Technol. 53,
    12586–12593.}

  Liess, M., Henz, S., Shahid, N., 2020. Modelling the synergistic effects
    of toxicant mixtures. Manuscript submitted for publication.
}
\usage{
multiple_stress
}
\description{
A data set of mixture toxicity experiments. Individuals of Daphnia magna were
exposed to combinations of food stress and multiple concentrations of the
pesticides esfenvalerate and prochloraz. The survival was recorded at 21 days
after contamination.
}
\keyword{datasets}
