% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mixture.R
\name{predict_mixture}
\alias{predict_mixture}
\title{Predict the effect of a mixture of two toxicants}
\usage{
predict_mixture(
  model_1,
  model_2,
  concentration_1,
  concentration_2,
  proportion_ca = 0.5
)
}
\arguments{
\item{model_1}{The ecxsys model of the toxicant that varies in concentration.}

\item{model_2}{The ecxsys model of the toxicant with a fixed concentration.}

\item{concentration_1}{The concentration of toxicant 1 in the mixture.}

\item{concentration_2}{The concentration of toxicant 2 in the mixture.}

\item{proportion_ca}{How much of the combined toxicant stress is determined
by concentration addition. Must be between 0 and 1.}
}
\value{
A vector of the effects of the mixture, scaled to the effect_max
  value of model_1.
}
\description{
This method is only suitable for experiments without environmental stress.
Any environmental stress in \code{model_1} or \code{model_2} is ignored.
}
\examples{
toxicant_model_1 <- ecxsys(
    concentration = c(0, 0.05, 0.5, 5, 30),
    hormesis_concentration = 0.5,
    effect_tox_observed = c(90, 81, 92, 28, 0),
)
toxicant_model_2 <- ecxsys(
    concentration = c(0, 0.1, 1, 10, 100, 1000),
    hormesis_concentration = 10,
    effect_tox_observed = c(26, 25, 24, 27, 5, 0),
    effect_max = 30
)
predict_mixture(
    toxicant_model_1,
    toxicant_model_2,
    c(0, 0.02, 0.2, 2, 20),
    3
)
}
