% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecxsys.R
\name{ecxsys}
\alias{ecxsys}
\title{ECx-SyS}
\usage{
ecxsys(
  concentration,
  hormesis_concentration,
  effect_tox_observed,
  effect_tox_env_observed = NULL,
  effect_max = 100,
  p = 3.2,
  q = 3.2
)
}
\arguments{
\item{concentration}{A vector of concentrations, one of which must be 0 to
indicate the control. Should be sorted in ascending order, otherwise it
will be sorted automatically.}

\item{hormesis_concentration}{The concentration where the hormesis occurs.
This is usually the concentration of the highest effect after the control.}

\item{effect_tox_observed}{A vector of effect values observed at the given
concentrations and in absence of environmental stress. Values must be
between 0 and \code{effect_max}.}

\item{effect_tox_env_observed}{Effect values observed in the presence of
environmental stress. Must be between 0 and \code{effect_max}.}

\item{effect_max}{The maximum value the effect could possibly reach. For
survival data in percent this should be 100 (the default).}

\item{p, q}{The shape parameters of the beta distribution. Default is 3.2.}
}
\value{
A list (of class ecxsys) containing many different objects with the
  most important being \code{curves}, a data frame containing effect and
  stress values at different concentrations. See \code{\link{predict_ecxsys}}
  for details.
}
\description{
The ECx-SyS model for modeling concentration-effect relationships which
indicate signs of hormesis.
}
\details{
It is advised to complete the curve down to zero for optimal prediction.
Therefore \code{effect_tox_observed} in the highest concentration should be
at or close to zero. If the model does not fit properly try adding an effect
of 0 at ten times the maximum observed concentration.

The vectors \code{concentration}, \code{effect_tox_observed} and
\code{effect_tox_env_observed} (if provided) must be of equal length and
sorted by increasing concentration.
}
\examples{
model <- ecxsys(
    concentration = c(0, 0.03, 0.3, 3, 10),
    hormesis_concentration = 0.3,
    effect_tox_observed = c(85, 76, 94, 35, 0),
    effect_tox_env_observed = c(24, 23, 32, 0, 0)
)

# Use effect_max if for example the effect is given as the number of
# surviving animals and the initial number of animals is 20:
model <- ecxsys(
    concentration = c(0, 0.03, 0.3, 3, 10),
    hormesis_concentration = 0.3,
    effect_tox_observed = c(17, 15.2, 18.8, 7, 0),
    effect_tox_env_observed = c(4.8, 4.6, 6.4, 0, 0),
    effect_max = 20
)

}
