% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecxsys.R
\name{ecxsys}
\alias{ecxsys}
\title{ECx-SyS}
\usage{
ecxsys(
  concentration,
  effect_tox_observed,
  effect_tox_env_observed,
  hormesis_concentration,
  hormesis_index,
  effect_max = 100,
  p = 3.2,
  q = 3.2
)
}
\arguments{
\item{concentration}{A vector of concentrations, one of which must be 0 to
indicate the control.}

\item{effect_tox_observed}{A vector of effect values observed at the given
concentrations and in absence of environmental stress. Values must be
between 0 and \code{effect_max}.}

\item{effect_tox_env_observed}{Effect values observed at the given
concentrations and in the presence of environmental stress. Values must be
between 0 and \code{effect_max}. This argument is optional and can be left
out to model without environmental stress.}

\item{hormesis_concentration}{The concentration where the hormesis is.}

\item{hormesis_index}{A single integer specifying the index of the hormesis
concentration in the concentration vector. This argument exists for
compatibility with older versions of this function. Use one of the the
hormesis arguments but not both.}

\item{effect_max}{The maximum value the effect could possibly reach. For
survival data in percent this should be 100 (the default).}

\item{p, q}{The shape parameters of the beta distribution. Default is 3.2.}
}
\value{
TODO: explain output list, especially the function fn
}
\description{
The ECx-SyS model.
}
\details{
TODO: improve this help text
}
\examples{
result <- ecxsys(
    concentration = c(0, 0.03, 0.3, 3, 10),
    effect_tox_observed = c(85, 76, 94, 35, 0),
    effect_tox_env_observed = c(24, 23, 32, 0, 0),
    hormesis_index = 3
)

}
