% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_tox.R
\name{multi_tox}
\alias{multi_tox}
\title{Predict the survival of binary toxicant mixtures}
\usage{
multi_tox(
  model_a,
  model_b,
  concentration_a,
  concentration_b,
  sa_contribution = 0.5,
  survival_max = 100
)
}
\arguments{
\item{model_a, model_b}{The ecxsys models of the toxicants.}

\item{concentration_a, concentration_b}{The concentrations of the toxicants in
the mixture. Both vectors must either be the same length or the longer
length must be a multiple of the shorter length. That's because the shorter
concentration vector gets recycled to the length of the longer one.}

\item{sa_contribution}{The proportion of stress addition contributing to the
calculation of the toxicant stress in the mixture. Must be between 0 and 1
where 1 stands for 100 \% stress addition.}

\item{survival_max}{Controls the scaling of the result. This represents the
maximum value the survival could possibly reach. For survival data in
percent this should be 100 (the default).}
}
\value{
A data frame with columns of the supplied concentrations and the
  corresponding mixture survival and stresses.
}
\description{
The Multi-TOX model predicts the effects of binary toxicant mixtures based on
three-phasic concentration-response relationships. See the publication for
details.
}
\details{
The predictions are symmetric, i.e. it does not matter which of the toxicant
models is 1 or 2 as long as the concentration arguments are supplied in the
right order. See the example below.

This method is only suitable for experiments without environmental stress.
Any environmental stress in \code{model_a} or \code{model_b} is ignored.
}
\examples{
toxicant_a  <- ecxsys(
    concentration = c(0, 0.05, 0.5, 5, 30),
    hormesis_concentration = 0.5,
    survival_tox_observed = c(90, 81, 92, 28, 0),
)
toxicant_b  <- ecxsys(
    concentration = c(0, 0.1, 1, 10, 100, 1000),
    hormesis_concentration = 10,
    survival_tox_observed = c(26, 25, 24, 27, 5, 0),
    survival_max = 30
)
multi_tox(
    toxicant_a ,
    toxicant_b ,
    c(0, 0.02, 0.2, 2, 20),
    c(0, 0.03, 0.4, 5, 10)
)

# Example of symmetric prediction:
conc_a <- c(0, 0.03, 0.3, 3)
conc_b <- 5.5
sa_contrib <- 0.75
mix_a <- multi_tox(toxicant_a , toxicant_b , conc_a, conc_b, sa_contrib)
mix_b <- multi_tox(toxicant_b , toxicant_a , conc_b, conc_a, sa_contrib)
identical(mix_a$survival, mix_b$survival)

}
