% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecxsys.R, R/plot_effect.R,
%   R/plot_system_stress.R
\name{plot_ecxsys}
\alias{plot_ecxsys}
\alias{plot_effect}
\alias{plot_system_stress}
\title{Plot the results of the ECx-SyS model}
\usage{
plot_effect(
  model,
  show_simple_model = FALSE,
  show_legend = FALSE,
  xlab = "concentration",
  ylab = "effect"
)

plot_system_stress(model, show_legend = FALSE)
}
\arguments{
\item{model}{The list returned from ecxsys().}

\item{show_simple_model}{Should the log-logistic models be plotted for
comparison? Defaults to \code{FALSE}.}

\item{show_legend}{Should the plot include a legend? Defaults to FALSE
because it may cover some points or lines depending on the plot size.}

\item{xlab, ylab}{Axis labels.}
}
\description{
Convenience functions to plot the observed and modeled effect and the
system stress with and without environmental stress.
}
\examples{
model <- ecxsys(
    effect_tox_observed = c(85, 76, 94, 35, 0),
    effect_tox_env_observed = c(24, 23, 32, 0, 0),
    concentration = c(0, 0.03, 0.3, 3, 10),
    hormesis_concentration = 0.3
)
plot_effect(model)
plot_system_stress(model)
}
