% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ecxsys.R
\name{predict_ecxsys}
\alias{predict_ecxsys}
\title{Predict ECxSyS at various concentrations}
\usage{
predict_ecxsys(model, concentration)
}
\arguments{
\item{model}{The output of a call to \code{\link{ecxsys}}.}

\item{concentration}{A numeric vector of concentrations.}
}
\value{
A data frame (of class "ecxsys_predicted") with the following
  columns:
  \describe{
    \item{concentration}{Concentrations regularly spaced on a logarithmic
    scale in the given concentration range. The control is approximated by
    the lowest non-control concentration times 1e-7.}
    \item{effect_tox_LL5}{The five-parameter log-logistic model of the
    effect derived from the observations under toxicant stress but without
    environmental stress.}
    \item{effect_tox}{Modeled effect resulting from toxicant and system
    stress.}
    \item{effect_tox_sys}{Modeled effect resulting from toxicant and system
    stress.}
    \item{stress_tox}{The toxicant stress.}
    \item{sys_tox}{System stress under toxicant stress conditions
    without environmental stress.}
    \item{stress_tox_sys}{The sum of \code{stress_tox} and
    \code{sys_tox}.}
    \item{effect_tox_env_LL5}{The five-parameter log-logistic model of the
    effect derived from the observations under toxicant stress with
    environmental stress.}
    \item{effect_tox_env}{Modeled effect resulting from toxicant and
    environmental stress.}
    \item{effect_tox_env_sys}{Modeled effect resulting from toxicant,
    environmental and system stress.}
    \item{stress_env}{Environmental stress.}
    \item{stress_tox_env}{The sum of toxicant and environmental stress.}
    \item{sys_tox_env}{System stress under toxicant and
    environmental stress conditions.}
    \item{stress_tox_env_sys}{The sum of \code{stress_tox_env} and
    \code{sys_tox_env}.}
  }
}
\description{
Predict ECxSyS at various concentrations
}
\examples{
model <- ecxsys(
    concentration = c(0, 0.03, 0.3, 3, 10),
    hormesis_concentration = 0.3,
    effect_tox_observed = c(85, 76, 94, 35, 0)
)
p <- predict_ecxsys(model, c(0.001, 0.01, 0.1, 1, 10))

}
