% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stressaddition-package.R
\docType{package}
\name{stressaddition-package}
\alias{stressaddition}
\alias{stressaddition-package}
\title{stressaddition: Modeling Tri-Phasic Concentration-Response Relationships}
\description{
The implementation of the tri-phasic concentration-response model 
    introduced in "Predicting low-concentration effects of pesticides" by 
    Liess, Henz and Knillmann (2019) <doi: 10.1038/s41598-019-51645-4>. It 
    allows modeling of ecotoxicological experiments where the response shows 
    signs of hormesis.
}
\details{
See the publication linked below for more information including
  equations.

  In the paper the model is introduced in the context of survival
  experiments. However, it can also be used to model other concentration
  dependent responses. For this reason the more general term "effect" instead
  of "survival" is used throughout the package.

  This package is not on CRAN. It is hosted on the UFZ GitLab server. Visit
  the repository linked below for the newest version. See the readme file
  there for instructions on how to install and update the package.
}
\references{
\href{https://doi.org/10.1038/s41598-019-51645-4}{Liess, M.,
  Henz, S. & Knillmann, S. Predicting low-concentration effects of
  pesticides. Sci Rep 9, 15248 (2019).}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.ufz.de/oekotox/stressaddition}
}

}
\author{
\strong{Maintainer}: Sebastian Henz \email{sebastian.henz@ufz.de} (\href{https://orcid.org/0000-0001-8299-8852}{ORCID})

Authors:
\itemize{
  \item Matthias Liess (\href{https://orcid.org/0000-0002-3321-8909}{ORCID})
}

Other contributors:
\itemize{
  \item Helmholtz-Zentrum fuer Umweltforschung GmbH - UFZ [copyright holder, funder]
}

}
