% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_effect_conversion.R
\name{Stressconversion}
\alias{Stressconversion}
\alias{effect_to_stress}
\alias{stress_to_effect}
\title{Convert Between Stress and Effect}
\usage{
effect_to_stress(effect, p = 3.2, q = 3.2)

stress_to_effect(stress, p = 3.2, q = 3.2)
}
\arguments{
\item{effect}{One or more effect values to convert to general stress.
Should be a value between 0 and 1. Values outside that interval are clipped
to the interval edges.}

\item{p, q}{The shape parameters of the beta distribution. Default is 3.2.}

\item{stress}{One or more stress values to convert to effect.}
}
\description{
Functions to convert effect to general stress or vice versa using the beta
distribution.
}
\details{
These are simple wrappers around the beta distribution function
\code{\link[stats:Beta]{pbeta}} and the beta quantile function
\code{\link[stats:Beta]{qbeta}}. \code{stress_to_effect} returns
\code{1 - pbeta(stress, p, q)}. \code{effect_to_stress} first clips the
effect to the interval [0, 1] and then returns
\code{qbeta(1 - effect, p, q)}.
}
\examples{
stress <- 0.3
effect <- stress_to_effect(stress)
effect_to_stress(effect)

}
